requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyyaml
httpx
openai
sse-starlette
selenium
playwright
aiofiles
markdown
aiomysql
mysql-connector-python
paramiko
python-jose[cryptography]
passlib[bcrypt]
chromadb
pyinstaller